Move instant launch data mocks into a separate file
